+++ init/restore-package-json.js
@@ -6,10 +6,16 @@ const { fgWhite } = require('./colors');
 
 function restorePackageJson() {
   try {
-    fs.unlinkSync('package.json');
-    fs.rmdirSync('node_modules ', { recursive: true });
+    if (fs.existsSync('package.json')) {
+      fs.unlinkSync('package.json');
+    }
+    if (fs.existsSync('node_modules')) {
+      fs.rmdirSync('node_modules ', { recursive: true });
+    }
     fs.renameSync('~package.json', 'package.json');
-    fs.unlinkSync('package-lock.json');
+    if (fs.existsSync('package-lock.json')) {
+      fs.unlinkSync('package-lock.json');
+    }
 
     console.info(fgWhite, 'Restoring package.json... Completed');
   } catch (error) {
